abstracted AppStepper, FilesHeader, & NetlifyBadge
