+++ src/components/application-stepper/NetlifyBadge.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+
+export default function NetlifyBadge() {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.netlifyBadge}>
+      <a href="https://www.netlify.com" target="_blank" rel="noopener noreferrer">
+        <img
+          src="https://www.netlify.com/img/global/badges/netlify-color-accent.svg"
+          alt="Deploys by Netlify"
+        />
+      </a>
+    </div>
+  );
+};
+
+const useStyles = makeStyles(theme => ({
+  netlifyBadge: {
+    width: '100%',
+    textAlign: 'center',
+    marginTop: theme.spacing(2),
+  },
+}));
